+++ src/routes/about.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router();
 
 router.get('/', (req, res) => {
-    res.send('MUA API (Matuco19 Utility API) is a multi-purpose API developed in Node.js.');
+    res.send('MUA API (Matuco19 Utility API) is a multi-purpose API developed in Node.js and maintained by Matuco19.');
 });
 
 module.exports = router;