+++ src/routes/index.js
@@ -3,7 +3,6 @@ const router = express.Router();
 
 router.get('/', (req, res) => {
     res.send('MUA API (Matuco19 Utility API) is a multi-purpose API developed in Node.js.');
-    // WIP for now
 });
 
 module.exports = router;