+++ CHANGELOG.md
@@ -5,5 +5,5 @@ Below, there are all the MUA API changelogs, from the most recent to the oldest,
 ## v0.1.0 - 12.xx.2025
 
 - Published as Open-Source in Github
-- Added 5 new endpoints: `/screenshot`, `/about`, `/analyze-sentiment`, `/cors-proxy`, and `/`
+- Added 6 new endpoints: `/screenshot`, `/about`, `/analyze-sentiment`, `/cors-proxy`, `/llm/token-counter`, and `/`
 - First Version